dynamic stats

+++ Assets/Scripts/Model/ProjectileSpec.cs
@@ -6,7 +6,7 @@ namespace DefaultNamespace.Model
     public struct ProjectileSpec
     {
         public int Damage;
-        public int Velocity;
+        public float Velocity;
         public int Lifespan;
         public float Radius;
     }
